final_edu_project/travel_service/requirements.txt
asgiref==3.6.0
Babel==2.12.1
backports.zoneinfo==0.2.1
Django==4.1.7
django-phonenumber-field==7.0.2
phonenumbers==8.13.8
psycopg2-binary==2.9.5
pytz==2023.3
sqlparse==0.4.3
